update to S2K prefix
