+++ kevm-pyk/src/kevm_pyk/utils.py
@@ -347,53 +347,3 @@ def legacy_explore(
                 id=id,
                 trace_rewrites=trace_rewrites,
             )
-
-
-def escaped_chars() -> list[str]:
-    return ['_', 'z']
-
-
-def escape_char(char: str) -> str:
-    return f'z{hex(ord(char)).removeprefix("0x")}'
-
-
-def unescape_seq(seq: str) -> str:
-    return chr(int(seq, base=16))
-
-
-def name_escaped(name: str, prefix: str = '') -> str:
-    escaped = prefix
-    iter_name = iter(name)
-    for char in iter_name:
-        if char in escaped_chars():
-            escaped += escape_char(char)
-        else:
-            escaped += char
-    return escaped
-
-
-def name_unescaped(name: str, prefix: str = '') -> str:
-    if not name.startswith(prefix):
-        raise RuntimeError(f'name {name} should start with {prefix}')
-    unescaped = name.removeprefix(prefix)
-    res = ''
-    unes_iter = iter(unescaped[:-1])
-    skipped = 0
-    for i, char in enumerate(unes_iter):
-        j = i + skipped
-        next_char = unescaped[j + 1]
-        if char == 'z':
-            res += unescape_seq(unescaped[(j + 1) : (j + 3)])
-            for _ in range(2):
-                try:
-                    next(unes_iter)
-                    skipped += 1
-                except StopIteration:
-                    # if we reached the end, write the last char and return
-                    break
-        else:
-            res += char
-        # write last char
-        if j + 2 == len(unescaped):
-            res += next_char
-    return res
